refactor(review): move appropriate methods to user and moderator services
