+++ apps/api/src/entities/review/review.controller.ts
@@ -3,12 +3,18 @@ import { Request, Response } from 'express';
 import { catchError, from, map, tap } from 'rxjs';
 import { Roles } from 'src/decorators/roles.decorator';
 import { RolesGuard } from 'src/guards/roles.guard';
+import { ModeratorService } from '../moderator/moderator.service';
 import { UserCookieData, UserRoles } from '../user/user.model';
+import { UserService } from '../user/user.service';
 import { ReviewService } from './review.service';
 
 @Controller('review')
 export class ReviewController {
-  constructor(private reviewService: ReviewService) { }
+  constructor(
+    private reviewService: ReviewService,
+    private moderatorService: ModeratorService,
+    private userService: UserService,
+  ) { }
 
   @UseGuards(RolesGuard)
   @Roles(UserRoles.MODERATOR)
@@ -16,7 +22,7 @@ export class ReviewController {
   async getArgumentAsModerator(@Req() req: Request, @Res() res: Response, @Param('ticketId') ticketId: string) {
     const user = (req as any).session.user as UserCookieData;
 
-    return from(this.reviewService.getArgumentAsModerator(user, ticketId))
+    return from(this.moderatorService.getArgumentAsModerator(user, ticketId))
       .pipe(
         tap(data => {
           if (!data) {
@@ -36,7 +42,7 @@ export class ReviewController {
   async getDebateAsModerator(@Req() req: Request, @Res() res: Response, @Param('ticketId') ticketId: string) {
     const user = (req as any).session.user as UserCookieData;
 
-    return from(this.reviewService.getDebateAsModerator(user, ticketId))
+    return from(this.moderatorService.getDebateAsModerator(user, ticketId))
       .pipe(
         tap(data => {
           if (!data) {
@@ -67,7 +73,7 @@ export class ReviewController {
   async getDebateAsUser(@Req() req: Request, @Res() res: Response, @Param('ticketId') ticketId: string) {
     const user = (req as any).session.user as UserCookieData;
 
-    return from(this.reviewService.getDebateAsUser(user, ticketId))
+    return from(this.userService.getDebateAsUser(user, ticketId))
       .pipe(
         tap(data => {
           if (!data) {
@@ -98,7 +104,7 @@ export class ReviewController {
   async getArgumentAsUser(@Req() req: Request, @Res() res: Response, @Param('ticketId') ticketId: string) {
     const user = (req as any).session.user as UserCookieData;
 
-    return from(this.reviewService.getArgumentAsUser(user, ticketId))
+    return from(this.userService.getArgumentAsUser(user, ticketId))
       .pipe(
         tap(data => {
           if (!data) {
