+++ apps/api/src/entities/user/user.service.ts
@@ -1,6 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { Pool } from 'pg';
 import { PG_PROVIDER_TOKEN } from '@decentdebates/db';
+import { ArgumentAsUser, DebateAsUser } from '../review/review.model';
 import { RegisterUserDTO } from './dtos/register-user.dto';
 import { User, UserCookieData, UserActivityArgument, UserActivityDebate } from './user.model';
 
@@ -146,6 +147,99 @@ export class UserService {
     } catch (err) {
       console.error(err);
       throw new Error('An error occurred while fetching the user\'s activity(arguments).');
+    } finally {
+      client.release();
+    }
+  }
+
+  async getDebateAsUser(user: UserCookieData, ticketId: string): Promise<DebateAsUser> {
+    const sqlStr = `
+      with debates_tags as (
+        select
+          d.id "debateId",
+          string_agg(dt.name, ',') "tags",
+          string_agg(dt.id::text, ',') "tagsIds"
+        from debate d
+        join assoc_debate_tag adt
+          on adt.debate_id = d.id
+        join debate_tag dt
+          on dt.id = adt.tag_id
+        group by d.id
+      )
+      select
+        d.id,
+        d.title,
+        d.ticket_id "ticketId",
+        d.created_at "createdAt",
+        d.modified_at "modifiedAt",
+        u.username "moderatorUsername",
+        u.id "moderatorId",
+        dt."tags",
+        t.board_list "boardList",
+        t.id "ticketId",
+        dt."tagsIds",
+        t.assigned_to "assignedToId",
+        t.created_by "userId"
+      from debate d
+      join debates_tags dt
+        on dt."debateId" = d.id
+      join ticket t
+        on d.ticket_id = t.id
+      left join "user" u
+        on u.id = t.assigned_to
+      where d.ticket_id = $1 and t.created_by = $2;
+    `;
+    const values = [ticketId, user.id];
+
+    const client = await this.pool.connect();
+    try {
+      const res = await client.query(sqlStr, values);
+      return res.rows[0];
+    } catch (err) {
+      console.log(err.message);
+      throw new Error('An error occurred while fetching the debate.');
+    } finally {
+      client.release();
+    }
+  }
+
+  async getArgumentAsUser(user: UserCookieData, ticketId: string): Promise<ArgumentAsUser> {
+    const sqlStr = `
+      select
+        t.id "ticketId",
+        u.id "moderatorId",
+        a.debate_id "debateId",
+        a.title "argumentTitle",
+        a.content "argumentContent",
+        a.counterargument_to "counterargumentToId",
+        a.type "argumentType",
+        d.title "debateTitle",
+        t.board_list "boardList",
+        aCounterarg.title "counterargumentToTitle",
+        u.username "moderatorUsername",
+        a.id "argumentId",
+        t.assigned_to "assignedToId",
+        t.created_by "userId"
+      from ticket t
+      join argument a
+        on a.ticket_id = t.id
+      join debate d
+        on d.id = a.debate_id
+      left join "user" u
+        on u.id = t.assigned_to
+      left join argument aCounterarg
+        on aCounterarg.id = a.counterargument_to
+      where t.id = $1 and a.is_draft = false and t.created_by = $2
+    `;
+    const values = [ticketId, user.id];
+
+    const client = await this.pool.connect();
+    try {
+      const res = await client.query(sqlStr, values);
+      return res.rows[0];
+    } catch (err) {
+      console.error(err.message);
+      throw err;
     } finally {
       client.release();
     }
